+++ packages/edit-site/src/components/page-pages/index.js
@@ -277,15 +277,19 @@ export default function PagePages() {
 		[ totalItems, totalPages ]
 	);
 
-	const { frontPageId, postsPageId, addNewLabel } = useSelect( ( select ) => {
-		const { getEntityRecord, getPostType } = select( coreStore );
-		const siteSettings = getEntityRecord( 'root', 'site' );
-		return {
-			frontPageId: siteSettings?.page_on_front,
-			postsPageId: siteSettings?.page_for_posts,
-			addNewLabel: getPostType( 'page' )?.labels?.add_new_item,
-		};
-	} );
+	const { frontPageId, postsPageId, addNewLabel, canCreatePage } = useSelect(
+		( select ) => {
+			const { getEntityRecord, getPostType, canUser } =
+				select( coreStore );
+			const siteSettings = getEntityRecord( 'root', 'site' );
+			return {
+				frontPageId: siteSettings?.page_on_front,
+				postsPageId: siteSettings?.page_for_posts,
+				addNewLabel: getPostType( 'page' )?.labels?.add_new_item,
+				canCreatePage: canUser( 'create', 'pages' ),
+			};
+		}
+	);
 
 	const fields = useMemo(
 		() => [
@@ -456,7 +460,10 @@ export default function PagePages() {
 		[ authors, view.type, frontPageId, postsPageId ]
 	);
 
-	const postTypeActions = usePostActions( 'page' );
+	const postTypeActions = usePostActions( {
+		postType: 'page',
+		context: 'list',
+	} );
 	const editAction = useEditPostAction();
 	const actions = useMemo(
 		() => [ editAction, ...postTypeActions ],
@@ -496,9 +503,14 @@ export default function PagePages() {
 		<Page
 			title={ __( 'Pages' ) }
 			actions={
-				addNewLabel && (
+				addNewLabel &&
+				canCreatePage && (
 					<>
-						<Button variant="primary" onClick={ openModal }>
+						<Button
+							variant="primary"
+							onClick={ openModal }
+							__next40pxDefaultSize
+						>
 							{ addNewLabel }
 						</Button>
 						{ showAddPageModal && (