+++ packages/editor/src/components/post-title/index.native.js
@@ -18,10 +18,66 @@ import { store as blockEditorStore, RichText } from '@wordpress/block-editor';
 import { store as editorStore } from '@wordpress/editor';
 import { __unstableStripHTML as stripHTML } from '@wordpress/dom';
 
+/** @typedef {import('./types').RichTextValue} RichTextValue */
+
 /**
  * Internal dependencies
  */
 import styles from './style.scss';
+
+/**
+ * Inserts content with title
+ *
+ * This function processes the given content and title, updating the title
+ * and content based on certain conditions. If the content is an array of
+ * blocks, it will check the first block for a heading or paragraph to use
+ * as the title. If the content is a string, it will strip HTML and update
+ * the title and content accordingly.
+ *
+ * @param {string}         title           The post title.
+ * @param {Array | string} content         The content to be processed. It can be an array of blocks or a string.
+ * @param {Function}       onUpdateTitle   Callback function to update the title.
+ * @param {Function}       onUpdateContent Callback function to update the content.
+ * @param {RichTextValue}  value           The initial value object, default is an object with empty text.
+ */
+export function insertContentWithTitle(
+	title,
+	content,
+	onUpdateTitle,
+	onUpdateContent,
+	value = create( { text: '' } )
+) {
+	if ( ! content.length ) {
+		return;
+	}
+
+	if ( typeof content !== 'string' ) {
+		const [ firstBlock ] = content;
+
+		if (
+			! title &&
+			( firstBlock.name === 'core/heading' ||
+				firstBlock.name === 'core/paragraph' )
+		) {
+			// Strip HTML to avoid unwanted HTML being added to the title.
+			// In the majority of cases it is assumed that HTML in the title
+			// is undesirable.
+			const contentNoHTML = stripHTML( firstBlock.attributes.content );
+			onUpdateTitle( contentNoHTML );
+			onUpdateContent( content.slice( 1 ) );
+		} else {
+			onUpdateContent( content );
+		}
+	} else {
+		// Strip HTML to avoid unwanted HTML being added to the title.
+		// In the majority of cases it is assumed that HTML in the title
+		// is undesirable.
+		const contentNoHTML = stripHTML( content );
+
+		const newValue = insert( value, create( { html: contentNoHTML } ) );
+		onUpdateTitle( toHTMLString( { value: newValue } ) );
+	}
+}
 
 class PostTitle extends Component {
 	constructor( props ) {
@@ -59,45 +115,24 @@ class PostTitle extends Component {
 	}
 
 	onPaste( { value, plainText, html } ) {
-		const { title, onInsertBlockAfter, onUpdate } = this.props;
+		const {
+			title,
+			onInsertBlockAfter: onInsertBlocks,
+			onUpdate,
+		} = this.props;
 
 		const content = pasteHandler( {
 			HTML: html,
 			plainText,
 		} );
 
-		if ( ! content.length ) {
-			return;
-		}
-
-		if ( typeof content !== 'string' ) {
-			const [ firstBlock ] = content;
-
-			if (
-				! title &&
-				( firstBlock.name === 'core/heading' ||
-					firstBlock.name === 'core/paragraph' )
-			) {
-				// Strip HTML to avoid unwanted HTML being added to the title.
-				// In the majority of cases it is assumed that HTML in the title
-				// is undesirable.
-				const contentNoHTML = stripHTML(
-					firstBlock.attributes.content
-				);
-				onUpdate( contentNoHTML );
-				onInsertBlockAfter( content.slice( 1 ) );
-			} else {
-				onInsertBlockAfter( content );
-			}
-		} else {
-			// Strip HTML to avoid unwanted HTML being added to the title.
-			// In the majority of cases it is assumed that HTML in the title
-			// is undesirable.
-			const contentNoHTML = stripHTML( content );
-
-			const newValue = insert( value, create( { html: contentNoHTML } ) );
-			onUpdate( toHTMLString( { value: newValue } ) );
-		}
+		insertContentWithTitle(
+			title,
+			content,
+			onUpdate,
+			onInsertBlocks,
+			value
+		);
 	}
 
 	setRef( richText ) {
@@ -166,7 +201,7 @@ class PostTitle extends Component {
 				<RichText.Raw
 					ref={ this.setRef }
 					accessibilityLabel={ this.getTitle( title, postType ) }
-					tagName={ 'p' }
+					tagName="p"
 					tagsToEliminate={ [ 'strong' ] }
 					unstableOnFocus={ this.props.onSelect }
 					onBlur={ this.props.onBlur } // Always assign onBlur as a props.
@@ -174,7 +209,7 @@ class PostTitle extends Component {
 					styles={ styles }
 					fontSize={ 24 }
 					lineHeight={ 1 }
-					fontWeight={ 'bold' }
+					fontWeight="bold"
 					deleteEnter
 					onChange={ ( value ) => {
 						this.props.onUpdate( value );